- smother scrolling with the keyboard and mouse scrolling - smother ship movement and rotation of the ship into the target direction

+++ engine/core/renderer.cpp
@@ -120,7 +120,7 @@ void Renderer::setRenderTarget(SDL_Texture* pTexture)
     }
 }
 
-double Renderer::getTimeDelta()
+float Renderer::getTimeDelta()
 {
     return delta;
 }
